Split Field implementation into multiple files
